+++ packages/utils/src/lib/reports/utils.ts
@@ -1,21 +1,12 @@
 import { InlineText, md } from 'build-md';
-import { join } from 'node:path';
 import {
+  AuditDiff,
   AuditReport,
   CategoryRef,
   IssueSeverity as CliIssueSeverity,
-  Format,
   Group,
   Issue,
-  PersistConfig,
-  Report,
-  reportSchema,
 } from '@code-pushup/models';
-import {
-  ensureDirectoryExists,
-  readJsonFile,
-  readTextFile,
-} from '../file-system';
 import { SCORE_COLOR_RANGE } from './constants';
 import { ScoredReport, SortableAuditReport, SortableGroup } from './types';
 
@@ -103,6 +94,28 @@ export function severityMarker(severity: 'info' | 'warning' | 'error'): string {
     return '⚠️';
   }
   return 'ℹ️';
+}
+
+export function formatScoreChange(diff: number): InlineText {
+  const marker = getDiffMarker(diff);
+  const text = formatDiffNumber(Math.round(diff * 1000) / 10); // round with max 1 decimal
+  return colorByScoreDiff(`${marker} ${text}`, diff);
+}
+
+export function formatValueChange({
+  values,
+  scores,
+}: Pick<AuditDiff, 'values' | 'scores'>): InlineText {
+  const marker = getDiffMarker(values.diff);
+  const percentage =
+    values.before === 0
+      ? values.diff > 0
+        ? Number.POSITIVE_INFINITY
+        : Number.NEGATIVE_INFINITY
+      : Math.round((100 * values.diff) / values.before);
+  // eslint-disable-next-line no-irregular-whitespace
+  const text = `${formatDiffNumber(percentage)} %`;
+  return colorByScoreDiff(`${marker} ${text}`, scores.diff);
 }
 
 export function calcDuration(start: number, stop?: number): number {
@@ -259,26 +272,6 @@ export function compareIssueSeverity(
     error: 2,
   };
   return levels[severity1] - levels[severity2];
-}
-
-type LoadedReportFormat<T extends Format> = T extends 'json' ? Report : string;
-
-export async function loadReport<T extends Format>(
-  options: Required<Omit<PersistConfig, 'format'>> & {
-    format: T;
-  },
-): Promise<LoadedReportFormat<T>> {
-  const { outputDir, filename, format } = options;
-  await ensureDirectoryExists(outputDir);
-  const filePath = join(outputDir, `${filename}.${format}`);
-
-  if (format === 'json') {
-    const content = await readJsonFile(filePath);
-    return reportSchema.parse(content) as LoadedReportFormat<T>;
-  }
-
-  const text = await readTextFile(filePath);
-  return text as LoadedReportFormat<T>;
 }
 
 export function throwIsNotPresentError(
