+++ packages/utils/src/lib/reports/utils.unit.test.ts
@@ -1,7 +1,5 @@
-import { vol } from 'memfs';
 import { describe, expect, it } from 'vitest';
-import { AuditReport, Issue, IssueSeverity, Report } from '@code-pushup/models';
-import { MEMFS_VOLUME, REPORT_MOCK, reportMock } from '@code-pushup/test-utils';
+import { AuditReport, Issue, IssueSeverity } from '@code-pushup/models';
 import { SCORE_COLOR_RANGE } from './constants';
 import { ScoredReport, SortableAuditReport, SortableGroup } from './types';
 import {
@@ -21,7 +19,6 @@ import {
   getSortableAuditByRef,
   getSortableGroupByRef,
   getSortedGroupAudits,
-  loadReport,
   scoreMarker,
   severityMarker,
 } from './utils';
@@ -364,65 +361,6 @@ describe('compareIssueSeverity', () => {
     expect([...severityArr].sort(compareIssueSeverity)).toEqual<
       IssueSeverity[]
     >(['info', 'warning', 'error']);
-  });
-});
-
-describe('loadReport', () => {
-  it('should load a valid JSON report', async () => {
-    vol.fromJSON(
-      {
-        [`report.json`]: JSON.stringify(reportMock()),
-        [`report.md`]: 'test-42',
-      },
-      MEMFS_VOLUME,
-    );
-
-    await expect(
-      loadReport({
-        outputDir: MEMFS_VOLUME,
-        filename: 'report',
-        format: 'json',
-      }),
-    ).resolves.toEqual(reportMock());
-  });
-
-  it('should load a markdown file', async () => {
-    vol.fromJSON(
-      {
-        [`report.dummy.md`]: 'test-7',
-        [`report.json`]: '{"test":42}',
-        [`report.md`]: 'test-42',
-      },
-      MEMFS_VOLUME,
-    );
-
-    await expect(
-      loadReport({
-        outputDir: MEMFS_VOLUME,
-        format: 'md',
-        filename: 'report',
-      }),
-    ).resolves.toBe('test-42');
-  });
-
-  it('should throw for an invalid JSON report', async () => {
-    vol.fromJSON(
-      {
-        [`report.json`]: JSON.stringify({
-          ...REPORT_MOCK,
-          plugins: [{ ...REPORT_MOCK.plugins[0]!, slug: '-Invalid_slug' }],
-        } satisfies Report),
-      },
-      MEMFS_VOLUME,
-    );
-
-    await expect(
-      loadReport({
-        outputDir: MEMFS_VOLUME,
-        filename: 'report',
-        format: 'json',
-      }),
-    ).rejects.toThrow('slug has to follow the pattern');
   });
 });
